requirements.txt
streamlit>=1.22
requests>=2.28
python-dotenv>=1.0
pandas>=2.0
google-generativeai>=0.2.0  # optional: only if you will call Gemini; keep if you plan to use it
pytest>=7.0
